Change code to adhere to PEP8 guidelines
Nothing functional was changed in the code, only line lengths and whitespace.
This was done to adhere to PEP8 guidelines, which is a requirement of the
project. Some changes are unnecessary according to me, where I had to split a
mathematical expression on 2 occasions in the code. However, I gave priority to
the project requirements and split the expressions over 2 lines.